# openapi.yaml
openapi: 3.0.3
info:
  title: Manufacturing Execution System API
  description: |
    Comprehensive MES API for jet engine component manufacturing.
    
    This API provides endpoints for managing work orders, quality control, 
    material tracking, equipment management, and comprehensive traceability 
    for aerospace manufacturing operations.
    
    ## Authentication
    All endpoints require JWT authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <token>
    ```
    
    ## Rate Limiting
    API requests are limited to 1000 requests per hour per user.
    
    ## Error Handling
    The API uses standard HTTP status codes. Error responses include:
    - `400` - Bad Request (validation errors)
    - `401` - Unauthorized (authentication required)
    - `403` - Forbidden (insufficient permissions)
    - `404` - Not Found
    - `409` - Conflict (business rule violation)
    - `422` - Unprocessable Entity (business validation error)
    - `500` - Internal Server Error
  version: 1.0.0
  contact:
    name: MES API Support
    email: support@mes-system.com
  license:
    name: Proprietary
servers:
  - url: https://api.mes-system.com/v1
    description: Production server
  - url: https://staging-api.mes-system.com/v1
    description: Staging server
  - url: http://localhost:3000/api/v1
    description: Development server
security:
  - BearerAuth: []
...
paths:
...
  /traceability/genealogy/{serialNumber}:
    get:
      tags:
        - Traceability
      summary: Get component genealogy
      description: Retrieve complete genealogy tree for a component
      parameters:
        - name: serialNumber
          in: path
          required: true
          schema:
            type: string
        - name: includeProcessData
          in: query
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Genealogy retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TraceabilityTree'
  /traceability/forward/{materialLot}:
    get:
      tags:
        - Traceability
      summary: Forward traceability
      description: Find all components produced from a material lot
      parameters:
        - name: materialLot
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Forward traceability retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ComponentTrace'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    TraceabilityTree:
      type: object
      properties:
        serialNumber:
          type: string
        partNumber:
          type: string
        partName:
          type: string
        manufacturedDate:
          type: string
          format: date
        workOrderNumber:
          type: string
        materials:
          type: array
          items:
            $ref: '#/components/schemas/MaterialTrace'
        operations:
          type: array
          items:
            $ref: '#/components/schemas/OperationTrace'
        children:
          type: array
          items:
            $ref: '#/components/schemas/TraceabilityTree'
    MaterialTrace:
      type: object
      properties:
        partNumber:
          type: string
        partName:
          type: string
        lotNumber:
          type: string
        serialNumber:
          type: string
        supplier:
          type: string
        certificateNumber:
          type: string
        heatNumber:
          type: string
        consumedDate:
          type: string
          format: date-time
        quantityConsumed:
          type: number
    OperationTrace:
      type: object
      properties:
        operationNumber:
          type: integer
        operationName:
          type: string
        workCenter:
          type: string
        equipment:
          type: string
        operator:
          type: string
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        processParameters:
          type: object
          additionalProperties: true
    ComponentTrace:
      type: object
      properties:
        serialNumber:
          type: string
        partNumber:
          type: string
        workOrderNumber:
          type: string
        manufacturedDate:
          type: string
          format: date
        customer:
          type: string
        status:
          type: string
...